requirements.txt
aiohttp==3.1.3
asn1crypto==0.24.0
astroid==1.6.0
async-timeout==2.0.1
attrs==17.4.0
autopep8==1.3.3
beautifulsoup4==4.5.3
bleach==1.5.0
build==1.0.2
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.1.4
disco-py==0.0.12
discord.py==1.0.0a1462+g60f0b22
discord.webhooks==0.0.3
earl-etf==2.1.2
gevent==1.2.2
gipc==0.6.0
gitdb2==2.0.2
GitPython==2.1.3
google-api-python-client==1.6.2
greenlet==0.4.13
holster==2.0.0
html5lib==0.9999999
httplib2==0.10.3
idna==2.6
idna-ssl==1.0.1
isort==4.2.15
lazy-object-proxy==1.3.1
lxml==3.8.0
Markdown==2.6.9
mccabe==0.6.1
multidict==4.2.0
nltk==3.2.4
numpy==1.13.1
oauth2client==4.1.2
olefile==0.44
pigar==0.7.1
Pillow==4.1.1
pluggy==0.6.0
pony==0.7.3
praw==5.4.0
prawcore==0.14.0
prettytable==0.7.2
protobuf==3.4.0
py==1.5.2
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycodestyle==2.3.1
pycparser==2.18
pyfiglet==0.7.5
pylint==1.8.1
PyNaCl==1.1.2
pyOpenSSL==17.5.0
PySocks==1.6.7
pytest==3.3.2
PythonGists==1.1.3
PyYAML==3.12
requests==2.18.4
rethinkdb==2.3.0.post6
rsa==3.4.2
six==1.11.0
smmap2==2.0.3
spice-api==1.0.4
strawpy==1.1
tensorflow==1.3.0
tensorflow-gpu==1.1.0
tensorflow-tensorboard==0.1.6
tqdm==4.15.0
ujson==1.35
update-checker==0.16
uritemplate==3.0.0
urllib3==1.22
websocket-client==0.44.0
websockets==3.4
Werkzeug==0.12.2
wrapt==1.10.11
XlsxWriter==1.0.2
yarl==1.2.3
